requirements.txt
python-telegram-bot
pyupbit
requests
